requirements.txt
langchain
langchain-openai
pyyaml
openai
streamlit